Multi-environments module format
- prebuild an UMD bundle on publish, in order to keep providing browser
and common.js assets
- convert an original source code to ES Module which is now a
primary entry point, so it can be used without a friction for consumers
which are written in ES Modules format, or use module bundlers like
Webpack or rollup
